Adding a convenience method for Rspec users
When testing against dynamic classes (AppConstant) Rspec will raise an
exception if you try to allow() or expect() on a class that does not
respond to the method. Rspec uses self.respond_to_missing? as an
underlying method to determine whether the stub is wrapping a real
method or not.

More information at
https://relishapp.com/rspec/rspec-mocks/docs/verifying-doubles/dynamic-classes